pyproject.toml
[project]
name = "career_alter_ego_chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.22.0",
    "openai>=1.68.2",
    "pypdf>=5.4.0",
    "python-dotenv>=1.0.1",
    "faiss-cpu>=1.7.4",
    "numpy>=1.26.4",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
